extract desktop scroller into ambiguous component
