+++ src/components/screens/carScroller/carScrollerDesktop.tsx
@@ -1,247 +1,21 @@
-import Image from 'next/image'
-import React, {
-  MouseEvent,
-  useCallback,
-  useEffect,
-  useMemo,
-  useRef,
-  useState,
-} from 'react'
-import { Block, Flex } from 'vcc-ui'
+import React from 'react'
 
 import { CarType } from '../../../types/cars'
-import { CAR_BLOCK_WIDTH_DESKTOP, CarBlock } from '../../organisms/carBlock'
+import { CarBlock } from '../../organisms/carBlock'
+import { DesktopScroller } from '../../organisms/desktopScroller'
 
 type Props = {
   cars: CarType[]
 }
 
 export const CarScrollerDesktop = ({ cars }: Props) => {
-  const [scrollIndex, setScrollIndex] = useState(1)
-  const [disableScrollLeft, setDisableScrollLeft] = useState(true)
-  const [disableScrollRight, setDisableScrollRight] = useState(false)
-
-  const carWrapperRef = useRef<HTMLDivElement>(null)
-
-  const carBlockSpace = 386
-  const carsInViewport = useMemo(
-    () => Math.floor(window.innerWidth / carBlockSpace),
-    []
-  )
-
-  const handleKeyScrollRight = (e: React.KeyboardEvent<HTMLImageElement>) => {
-    if (e.key === ' ') {
-      handleScrollRight()
-    }
-  }
-
-  const handleKeyScrollLeft = (e: React.KeyboardEvent<HTMLImageElement>) => {
-    if (e.key === ' ') {
-      handleScrollLeft()
-    }
-  }
-
-  const handleScrollRight = () => {
-    const element = carWrapperRef.current
-    const offset = 15
-    const carBlockSpace = CAR_BLOCK_WIDTH_DESKTOP - offset
-    const nextRightScroll = carBlockSpace * scrollIndex
-    if (!element || nextRightScroll !== element.scrollLeft + carBlockSpace) {
-      return
-    }
-    if (disableScrollLeft) {
-      setDisableScrollLeft(false)
-    }
-    if (nextRightScroll >= element.scrollWidth - window.innerWidth) {
-      if (cars.length - carsInViewport === scrollIndex) {
-        element.scrollTo({
-          left: element.scrollWidth,
-          behavior: 'smooth',
-        })
-        setScrollIndex((oldIndex) => oldIndex + 1)
-      }
-      setDisableScrollRight(true)
-      return
-    }
-    element.scrollTo({
-      left: nextRightScroll,
-      behavior: 'smooth',
-    })
-
-    setScrollIndex((oldIndex) => oldIndex + 1)
-  }
-
-  const handleScrollLeft = () => {
-    const element = carWrapperRef.current
-    const offset = 15
-    const carBlockSpace = CAR_BLOCK_WIDTH_DESKTOP - offset
-    const nextLeftScroll = carBlockSpace * (scrollIndex - 1) - carBlockSpace
-    if (
-      !element ||
-      (!(
-        nextLeftScroll >=
-        element.scrollLeft - (nextLeftScroll + carBlockSpace)
-      ) &&
-        scrollIndex === 1)
-    ) {
-      return
-    }
-    if (disableScrollRight) {
-      setDisableScrollRight(false)
-    }
-    if (nextLeftScroll <= 0) {
-      if (element.scrollLeft !== 0) {
-        element.scrollTo({
-          left: 0,
-          behavior: 'smooth',
-        })
-        setScrollIndex((oldIndex) => oldIndex - 1)
-      }
-      setDisableScrollLeft(true)
-      return
-    }
-    element.scrollTo({
-      left: nextLeftScroll,
-      behavior: 'smooth',
-    })
-    setScrollIndex((oldIndex) => oldIndex - 1)
-  }
-
-  const [touchStart, setTouchStart] = useState<number | null>(null)
-  const [touchEnd, setTouchEnd] = useState<number | null>(null)
-
-  const minSwipeDistance = 50
-
-  const onTouchStart = (e: MouseEvent<HTMLDivElement>) => {
-    setTouchEnd(null)
-    setTouchStart(e.clientX)
-  }
-
-  const onTouchMove = (e: MouseEvent<HTMLDivElement>) => {
-    setTouchEnd(e.clientX)
-  }
-
-  const onTouchEnd = () => {
-    if (!touchStart || !touchEnd) return
-    const distance = touchStart - touchEnd
-    const isLeftSwipe = distance > minSwipeDistance
-    const isRightSwipe = distance < -minSwipeDistance
-    // @todo better logic for detecting longer or faster swipe distances
-    if (isLeftSwipe) {
-      handleScrollRight()
-    } else if (isRightSwipe) {
-      handleScrollLeft()
-    }
-  }
-
-  useEffect(() => {
-    const element = carWrapperRef.current
-    if (!element) {
-      return
-    }
-    const allCarsVisible = carsInViewport >= cars.length
-    setScrollIndex(1)
-    setDisableScrollRight(allCarsVisible)
-    setDisableScrollLeft(true)
-    element.scrollTo({ left: 0, behavior: 'smooth' })
-  }, [cars, carsInViewport])
-
-  const detectTabKey = useCallback(
-    (e: KeyboardEvent) => {
-      const element = carWrapperRef.current
-      if (!element) {
-        return
-      }
-      if (e.key == 'Tab') {
-        if (element.scrollLeft > 0) {
-          setDisableScrollLeft(false)
-        }
-
-        if (document.activeElement?.className === 'car-block') {
-          const index = cars.findIndex(
-            (car) => car.id === document.activeElement?.id
-          )
-          const nextScrollIndex = index + 2 - carsInViewport
-          setScrollIndex(nextScrollIndex < 1 ? 1 : nextScrollIndex)
-          document.activeElement?.scrollIntoView({ behavior: 'smooth' })
-          if (index + 1 === cars.length) {
-            setDisableScrollRight(true)
-          }
-        }
-      }
-      if (e.shiftKey && e.key === 'Tab') {
-        if (document.activeElement?.className === 'car-block') {
-          const index = cars.findIndex(
-            (car) => car.id === document.activeElement?.id
-          )
-          const nextScrollIndex = index + 2 - carsInViewport
-          setScrollIndex(nextScrollIndex < 1 ? 1 : nextScrollIndex)
-          document.activeElement?.scrollIntoView({ behavior: 'smooth' })
-          if (index + 1 === 1) {
-            setDisableScrollLeft(true)
-          }
-          if (index + 1 < cars.length) {
-            setDisableScrollRight(false)
-          }
-        }
-      }
-    },
-    [cars, carsInViewport]
-  )
-
-  useEffect(() => {
-    document.addEventListener('keyup', detectTabKey)
-
-    return () => document.removeEventListener('keyup', detectTabKey)
-  }, [detectTabKey])
-
   return (
     <>
-      <Flex
-        ref={carWrapperRef}
-        style={{
-          flexDirection: 'row',
-          overflow: 'hidden',
-          whiteSpace: 'nowrap',
-          margin: '0 auto',
-          width: carBlockSpace * carsInViewport,
-        }}
-        onMouseDown={onTouchStart}
-        onMouseMove={onTouchMove}
-        onMouseUp={onTouchEnd}
-      >
+      <DesktopScroller itemBlockSpace={386}>
         {cars.map((car) => (
           <CarBlock key={car.id} car={car} />
         ))}
-      </Flex>
-      <Block style={{ position: 'absolute', bottom: 20, right: 20 }}>
-        <Image
-          src='/chevron-circled.svg'
-          width={40}
-          height={40}
-          style={{
-            transform: 'rotate(180deg)',
-            cursor: 'pointer',
-            opacity: disableScrollLeft ? 0.6 : 1,
-          }}
-          alt='scroll-left'
-          onClick={disableScrollLeft ? undefined : handleScrollLeft}
-          onKeyDown={disableScrollLeft ? undefined : handleKeyScrollLeft}
-          tabIndex={0}
-          role='button'
-        />
-        <Image
-          src='/chevron-circled.svg'
-          width={40}
-          height={40}
-          style={{ cursor: 'pointer', opacity: disableScrollRight ? 0.6 : 1 }}
-          alt='scroll-right'
-          onClick={disableScrollRight ? undefined : handleScrollRight}
-          onKeyDown={disableScrollRight ? undefined : handleKeyScrollRight}
-          tabIndex={0}
-          role='button'
-        />
-      </Block>
+      </DesktopScroller>
     </>
   )
 }
